pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plex-mcp-advanced"
version = "2.0.0"
description = "FastMCP 2.14.3+ & MCPB package for Plex Media Server management with conversational AI and sampling"
authors = [
    {name = "Sandra Schipal", email = "sandra@sandraschi.dev"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["mcp", "plex", "media", "streaming", "fastmcp", "glama-gold", "mcpb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]

dependencies = [
    "fastmcp>=2.14.3,<3.0.0",
    "aiohttp>=3.9.0,<4.0.0",
    "aiofiles>=23.2.0,<24.0.0",
    "plexapi>=4.15.0,<5.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pyyaml>=6.0.0,<7.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "black>=24.0.0,<25.0.0",
    "isort>=5.12.0,<6.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0,<2.0.0",
    "bandit>=1.7.0,<2.0.0",
    "safety>=2.3.0,<3.0.0",
    "types-PyYAML>=6.0.0,<7.0.0",
    "pre-commit>=3.6.0"
]

[project.scripts]
plex-mcp-advanced = "plex_mcp.server:main"

[project.urls]
Homepage = "https://github.com/yourusername/plex-mcp"
Repository = "https://github.com/yourusername/plex-mcp.git"
Documentation = "https://github.com/yourusername/plex-mcp#readme"
Changelog = "https://github.com/yourusername/plex-mcp/blob/main/CHANGELOG.md"
Issues = "https://github.com/yourusername/plex-mcp/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "S"]
ignore = ["E501", "B008", "C901", "B904"]
exclude = ["__pycache__", ".git", ".venv", "venv", "dist", "build"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # Allow assert in tests

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

